Fix viewer presence tracking (#986)
Supabase realtime hates React re-renders - we were trying to join and
leave the same channel a bunch of times which messed it up and caused
race conditions. Updated presence to use a similar method to realtime
events, where we store the connection separately and track a listener
function per component.